Move `getChannelById` to util.js
